chore: setup the environment
Setup the environment for testing and to leverage a component library
for expedience of building the site. It should, also, be noted that MUI
can only be used when a component is rendered client-side, instead of,
the prefered server-side. This is issue accures as of Next.js 13, and is
explained in further detail:
mui/material-ui#34898, mui/material-ui#37315