refactor (database): change part_of_speech field
Change the part_of_speech field in the part_of_speech table to
full_form. This is to better reflect the dichotomy between abbreviation
and the full form representation of the part of speech. This, too,
removes the confusing redundancy when refrencing this data. Sort of
like, defining a word using its own root. This leads to circular logic
and must be avoided.