fix: prevent cypress from crashing
Cypress crashed when restarting tests even when all tests would
successfully pass. The issue seemed to stem from the connection closing.
Then, there would be no way for the connection to re-open for Cypress
to use. This seems  to be a conterintuitive solution to a possible non-
problem as now the production code has been changed to help the tests,
instead of, the sole purpose of raising confidence in the necessary
production code.